# openapi/update-aggregator.yaml
openapi: 3.0.3
info:
  title: update-aggregator API
  description: >
    Central service update-detector agents push their status to. New agents
    are held as pending until approved via /admin; approved agents' data
    powers the /widgets/* endpoints for a Homepage dashboard.
    Also served live at GET /openapi.yaml on the running aggregator.
  version: "1.0"

security: []

servers:
  - url: http://{host}:{port}
    variables:
      host:
        default: localhost
      port:
        default: "8080"

paths:
  /enroll:
    post:
      summary: Agent self-registration (trust-on-first-contact)
      description: >
        Unauthenticated — an agent has nothing to prove on its very first
        contact. Creates a pending record for an unknown agent_id; for a
        known agent_id, confirms the token matches (idempotent re-announce)
        or reports a conflict if it doesn't.
      operationId: enroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollRequest"
      responses:
        "200":
          description: Current approval status for this agent_id
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, approved, rejected]
        "400":
          description: Missing agent_id, hostname, or token
        "409":
          description: agent_id already registered with a different token

  /report:
    post:
      summary: Push a detection result
      description: >
        Accepted only for an approved agent presenting its correct token.
        Called after every local check cycle; failures should be logged by
        the agent, not treated as fatal — local detection/Gatus/Telegram
        must keep working regardless of the aggregator's reachability.
      operationId: report
      security:
        - agentToken: []
      parameters:
        - name: X-Agent-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: >
          Same Status shape update-detector serves at its own GET /status
          (see update-detector.yaml, Status schema, in this same directory).
        content:
          application/json:
            schema:
              $ref: "update-detector.yaml#/components/schemas/Status"
      responses:
        "200":
          description: Report accepted
        "401":
          description: Unknown agent or invalid token
        "403":
          description: Agent known but not (yet) approved

  /companion/stream:
    get:
      summary: SSE stream of apply Actions for a host-native companion
      description: >
        Long-lived Server-Sent-Events response (plain HTTP/1.1, one Action
        at a time, plus periodic ": heartbeat" comments). A companion
        process holds this connection open and reconnects with backoff if
        it drops; being connected here is what makes POST
        /admin/agents/{id}/apply succeed for that agent instead of 409ing.
      operationId: companionStream
      security:
        - agentToken: []
      parameters:
        - name: X-Agent-ID
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: text/event-stream of Action payloads
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/Action"
        "401":
          description: Unknown agent or invalid token
        "403":
          description: Agent known but not (yet) approved

  /companion/result:
    post:
      summary: Report the outcome of a previously pushed Action
      description: >
        Recorded in a small in-memory per-agent action log (not persisted)
        and, if a notifier is configured, raises an alert (e.g. "update
        applied on web01").
      operationId: companionResult
      security:
        - agentToken: []
      parameters:
        - name: X-Agent-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionResult"
      responses:
        "200":
          description: Result recorded
        "400":
          description: Missing action_id
        "401":
          description: Unknown agent or invalid token
        "403":
          description: Agent known but not (yet) approved

  /admin:
    get:
      summary: Approval dashboard (HTML)
      description: >
        Lists agents by Pending/Approved/Rejected with Approve/Reject/Revoke
        form actions. No authentication of its own — keep on a private
        network or behind your own reverse-proxy auth.
      operationId: getAdminPage
      responses:
        "200":
          description: HTML page
          content:
            text/html:
              schema:
                type: string

  /admin/agents/{id}/approve:
    post:
      summary: Approve a pending or rejected agent
      operationId: approveAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "303":
          description: Redirect back to /admin
        "404":
          description: Unknown agent id

  /admin/agents/{id}/reject:
    post:
      summary: Reject a pending agent, or revoke an approved one
      operationId: rejectAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "303":
          description: Redirect back to /admin
        "404":
          description: Unknown agent id

  /admin/agents/{id}/apply:
    post:
      summary: Trigger a package upgrade on one agent's host
      description: >
        Human-triggered, and gated by a shared secret (ADMIN_APPLY_SHARED_SECRET)
        checked independently of whatever reverse-proxy auth fronts this
        endpoint -- disabled entirely (501) until that secret is configured.
        Requires a companion currently connected via GET /companion/stream
        for this agent, and the companion independently re-validates any
        named packages against its own host's last-known pending upgrades
        before acting -- this endpoint alone does not grant arbitrary
        command execution even if the aggregator itself were compromised.
      operationId: applyAgentAction
      security:
        - adminApplySecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRequest"
      responses:
        "202":
          description: Action accepted and pushed to the connected companion
          content:
            application/json:
              schema:
                type: object
                properties:
                  action_id:
                    type: string
        "400":
          description: Invalid type, or packages missing for type=packages
        "403":
          description: Missing or invalid X-Admin-Apply-Secret
        "404":
          description: Unknown agent id, or agent not approved
        "409":
          description: No companion currently connected for this agent
        "501":
          description: Apply is disabled (ADMIN_APPLY_SHARED_SECRET not configured)

  /widgets/summary:
    get:
      summary: Fleet-wide totals for a Homepage "Custom API" widget
      description: Computed across approved agents only.
      operationId: getWidgetSummary
      responses:
        "200":
          description: Fleet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"

  /widgets/hosts:
    get:
      summary: Per-host summaries for approved agents
      operationId: getWidgetHosts
      responses:
        "200":
          description: List of approved hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostSummary"

  /widgets/hosts/{hostname}:
    get:
      summary: One host's latest full report
      description: >
        Same JSON shape as an agent's own GET /status, so a Homepage widget
        config is identical whether it targets the aggregator or the agent
        directly. Looks up the most-recently-seen approved agent claiming
        this hostname.
      operationId: getWidgetHost
      parameters:
        - name: hostname
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The host's latest reported Status
          content:
            application/json:
              schema:
                $ref: "update-detector.yaml#/components/schemas/Status"
        "404":
          description: No approved agent with a report for this hostname yet

  /widgets/packages:
    get:
      summary: Fleet-wide list of pending package upgrades
      description: >
        Flattens every approved, reporting agent's pending package upgrades
        into one list, so a single Homepage widget (or any other consumer)
        can show "what needs updating" across the whole fleet without
        querying each host individually.
      operationId: getWidgetPackages
      parameters:
        - name: security
          in: query
          required: false
          description: If "true", only include packages flagged as security updates.
          schema:
            type: string
            enum: ["true"]
      responses:
        "200":
          description: Pending packages across every approved host
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingPackage"

  /openapi.yaml:
    get:
      summary: This OpenAPI spec
      operationId: getOpenAPISpec
      responses:
        "200":
          description: OpenAPI 3.0 spec for this service
          content:
            application/yaml:
              schema:
                type: string

components:
  securitySchemes:
    agentToken:
      type: http
      scheme: bearer
      description: The agent's self-generated token, paired with its X-Agent-ID header.
    adminApplySecret:
      type: apiKey
      in: header
      name: X-Admin-Apply-Secret
      description: >
        Shared secret configured via ADMIN_APPLY_SHARED_SECRET. Intended to
        be injected by a reverse-proxy (e.g. Authentik) after successful
        human auth; the aggregator checks it independently either way.

  schemas:
    Action:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [packages, upgrade, full-upgrade]
        packages:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

    ActionResult:
      type: object
      required: [action_id]
      properties:
        action_id:
          type: string
        success:
          type: boolean
        message:
          type: string

    ApplyRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [packages, upgrade, full-upgrade]
        packages:
          type: array
          items:
            type: string
          description: Required (and non-empty) when type=packages; ignored otherwise.

    EnrollRequest:
      type: object
      required: [agent_id, hostname, token]
      properties:
        agent_id:
          type: string
          description: Self-generated by the agent on first run (crypto/rand), persisted locally.
        hostname:
          type: string
          example: web01
        token:
          type: string
          description: Self-generated by the agent alongside agent_id. Stored server-side only as a sha256 hash.

    SummaryResponse:
      type: object
      properties:
        hosts_total:
          type: integer
        hosts_reporting:
          type: integer
          description: Approved hosts that have sent at least one report.
        hosts_ok:
          type: integer
        hosts_reboot_required:
          type: integer
        hosts_os_upgrade_available:
          type: integer
        packages_upgradable_total:
          type: integer
        packages_upgradable_security:
          type: integer
        generated_at:
          type: string
          format: date-time

    HostSummary:
      type: object
      properties:
        hostname:
          type: string
        agent_id:
          type: string
        has_report:
          type: boolean
        ok:
          type: boolean
        last_seen:
          type: string
          format: date-time
        upgradable_total:
          type: integer
        upgradable_security:
          type: integer
        reboot_required:
          type: boolean
        os_update_available:
          type: boolean

    PendingPackage:
      type: object
      required: [hostname, name, candidate_version]
      properties:
        hostname:
          type: string
          example: web01
        name:
          type: string
          example: curl
        current_version:
          type: string
          example: 7.81.0-1ubuntu1.15
        candidate_version:
          type: string
          example: 7.81.0-1ubuntu1.16
        security:
          type: boolean